+++ Hydra_Example/src/.ipynb_checkpoints/utils-checkpoint.py
@@ -1,6 +1,31 @@
+from omegaconf import DictConfig, OmegaConf
+import hydra
 
+import os
+import sys
+import runpy
+import numpy as np
+from functools import partial
+from numbers import Number
+import warnings
+
+import sirf.STIR as pet
+pet.set_verbosity(0)
+import sirf.Reg as reg
+from cil.framework import BlockDataContainer, ImageGeometry, BlockGeometry
+from cil.optimisation.algorithms import PDHG, SPDHG
+from cil.optimisation.functions import \
+    KullbackLeibler, BlockFunction, IndicatorBox, MixedL21Norm, ScaledFunction, TotalVariation
+from cil.optimisation.operators import \
+    CompositionOperator, BlockOperator, LinearOperator, GradientOperator, ScaledOperator
+from cil.plugins.ccpi_regularisation.functions import FGP_TV
+from ccpi.filters import 
+
+cil_path = '/home/jovyan/Hackathon-000-Stochastic-Algorithms/cil/'
+sys.path.append(cil_path)
 
 
 
 def set_up_acquisition_model_with_data(acquisition_model, dataset):
-    print("Not ready")
+    acquisition_model.set_acquisition_sensitivity(dataset.multiplicative_factors)
+    acquisition_model.set_up(dataset.acquisition_data, dataset.image_template)
