+++ Hydra_Example/src/Classes/.ipynb_checkpoints/Prior-checkpoint.py
@@ -24,24 +24,9 @@ from ccpi.filters import
 cil_path = '/home/jovyan/Hackathon-000-Stochastic-Algorithms/cil/'
 sys.path.append(cil_path)
 
-from sirf.Utilities import examples_data_path
-
-class Dataset(object):
+class Prior(object):
     def __init__(self,cfg):
         self.cfg=cfg
-        
-        if cfg.dataset.name = 'test':
-            self.acq_data = pet.ImageData('{}/{}/{}'.format(
-                examples_data_path("PET"),
-                'thorax_single_slice',
-                'emission.hv')
-            self.multiplicative_factor = pet.ImageData('{}/{}/{}'.format(
-                examples_data_path("PET"),
-                'thorax_single_slice',
-                'attenuation.hv')
-            self.additive_factor = self.acq_data.clone()
-            self.additive_factor.fill(0.01*self.acq_data.to_numpy())
-                                                       
-        self.image_template = self.acq_data.create_uniform_image(0.0)
-                                    
-            
+    
+        if self.cfg.functionals.prior.name == "NonNeg"
+                self.prior = IndicatorBox(lower=0)
