+++ Hydra_Example/src/Factories/.ipynb_checkpoints/DataFitFactory-checkpoint.py
@@ -14,8 +14,7 @@ pet.set_verbosity(0)
 import sirf.Reg as reg
 from cil.framework import BlockDataContainer, ImageGeometry, BlockGeometry
 from cil.optimisation.algorithms import PDHG, SPDHG
-from cil.optimisation.functions import \
-    KullbackLeibler, BlockFunction, IndicatorBox, MixedL21Norm, ScaledFunction, TotalVariation
+from cil.optimisation.functions import KullbackLeibler, OperatorCompositionFunction
 from cil.optimisation.operators import \
     CompositionOperator, BlockOperator, LinearOperator, GradientOperator, ScaledOperator
 from cil.plugins.ccpi_regularisation.functions import FGP_TV
@@ -24,10 +23,13 @@ from ccpi.filters import
 cil_path = '/home/jovyan/Hackathon-000-Stochastic-Algorithms/cil/'
 sys.path.append(cil_path)
 
-class AcquisitionModel(object):
+class DataFitFactory(object):
     def __init__(self,cfg):
         self.cfg=cfg
+                
+    def set_up(self,dataset,acquisition_model):
+        if self.cfg.acq_model.num_subsets == 1 and self.cfg.functionals.datafit.KL:
+            f = KullbackLeibler(b=dataset.acq_data, eta=dataset.additive_factor)
+            self.datafit = OperatorCompositionFunction(f,acquisition_model)
         
-        if self.cfg.acq_model.num_subsets == 1:
-            self.acquisition_model = pet.AcquisitionModelUsingRayTracingMatrix()
-            self.acquisition_model.set_num_tangential_LORs(self.cfg.acq_model.lor)
+
