--- Hydra_Example/src/Factories/.ipynb_checkpoints/FunctionalsFactory-checkpoint.py
@@ -1,4 +1,0 @@
-
-class FunctionalsFactory(object):
-    def __init__(self,acquisition_model, dataset):
-        self.cfg=cfg
