--- Hydra_Example/src/Factories/AlgorithmsFactory.py
@@ -1,6 +1,0 @@
-
-
-class AlgorithmFactory(object):
-    def __init__(self,functionals,acquisition_model,quality_metrics, warm_start_image):
-        self.cfg=cfg
-
