--- Hydra_Example/src/Factories/FunctionalsFactory.py
@@ -1,4 +1,0 @@
-
-class FunctionalsFactory(object):
-    def __init__(self,acquisition_model, dataset):
-        self.cfg=cfg
